google-vertexai-rag/requirements.txt
google-cloud-aiplatform
vertexai
pdfplumber
python-docx
flask
flask-cors
werkzeug

# 高效向量搜索依赖
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0

# 可选：GPU加速版本（如果有CUDA支持）
# faiss-gpu>=1.7.4  # 替换faiss-cpu以获得GPU加速

# 数据处理和缓存
scikit-learn>=1.3.0
pandas>=2.0.0 